Move more time-related parser code into time module
